Refactored HomeInsurance.java to use int instead of long for the fields 'homeAmount' and 'contentsAmount'.

+++ src/baminsurances/data/HolidayHomeInsurance.java
@@ -29,7 +29,7 @@ public class HolidayHomeInsurance extends HomeInsurance {
     public HolidayHomeInsurance(Employee employee, int premium, int amount,
             String terms, String streetAddress, String zipCode,
             HomeType homeType, String buildingMaterial, String standard,
-            int squareMetres, long homeAmount, long contentsAmount,
+            int squareMetres, int homeAmount, int contentsAmount,
             boolean rentedOut) {
         super(employee, premium, amount, terms, streetAddress, zipCode,
                 homeType, buildingMaterial, standard, squareMetres, homeAmount,
